requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
h11==0.16.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.6.3
numpy==2.0.2
pillow==11.3.0
propcache==0.3.2
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.1
tabulate==0.9.0
together==1.4.6
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
